pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "structa-server"
version = "0.1.0"
description = "A project management server with GraphQL, Redis, and Flask."
authors = [
    { name = "Gabriel Santos", email = "grs_santos2019@outlook.com" }
]
readme = "README.md"
repository = "https://github.com/Grs2080w/structa-server.git"
dependencies = [
    "anyio>=4.9.0",
    "ariadne>=0.26.1",
    "bcrypt>=4.3.0",
    "blinker>=1.9.0",
    "cffi>=1.17.1",
    "click>=8.1.8",
    "colorama>=0.4.6",
    "cryptography>=44.0.2",
    "defusedxml>=0.7.1",
    "Flask>=3.1.0",
    "flask-cors>=5.0.1",
    "fonttools>=4.56.0",
    "graphql-core>=3.2.5",
    "idna>=3.10",
    "itsdangerous>=2.2.0",
    "Jinja2>=3.1.6",
    "MarkupSafe>=3.0.2",
    "mistletoe>=1.4.0",
    "pillow>=11.1.0",
    "pycparser>=2.22",
    "PyJWT>=2.10.1",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "sniffio>=1.3.1",
    "starlette>=0.46.1",
    "typing_extensions>=4.13.0",
    "Werkzeug>=3.1.3",
    "fpdf2 @ git+https://github.com/PyFPDF/fpdf2.git@6f58e68a2109b66b270b4488af9d60cacdb0cd1a"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["flaskr", "config", "otp", "cryptography"]

[tool.setuptools.package-data]
"flaskr" = ["graphql/schema.graphql"]
"config" = [".env"]
"otp" = ["*.py"]
"cryptography" = ["*.py"]

[project.scripts]
start = "flask --app src/flaskr run"
